Small fixes to Training python examples for 4.4. See #9365

+++ examples/Training/python/Create_Image.py
@@ -31,10 +31,10 @@ imageId = 27544     # This image must have at least 2 channels
 # This example demonstrates the usage of the convenience method
 # createImageFromNumpySeq() Here we create a multi-dimensional image from a
 # hard-coded array of data.
-from numpy import array
+from numpy import array, int8
 sizeX, sizeY, sizeZ, sizeC, sizeT = 5, 4, 1, 2, 1
-plane1 = array([[0, 1, 2, 3, 4], [5, 6, 7, 8, 9], [0, 1, 2, 3, 4], [5, 6, 7, 8, 9]])
-plane2 = array([[5, 6, 7, 8, 9], [0, 1, 2, 3, 4], [5, 6, 7, 8, 9], [0, 1, 2, 3, 4]])
+plane1 = array([[0, 1, 2, 3, 4], [5, 6, 7, 8, 9], [0, 1, 2, 3, 4], [5, 6, 7, 8, 9]], dtype=int8)
+plane2 = array([[5, 6, 7, 8, 9], [0, 1, 2, 3, 4], [5, 6, 7, 8, 9], [0, 1, 2, 3, 4]], dtype=int8)
 planes = [plane1, plane2]
